feat: separate auth into backends
Problem: we have a lot of trouble with custom auth needs.
There is not one solution that works for everyone.
Solution: start a design that separates auth into modules.
Likely this will need a lot of testing, but this should
be a start.
