+++ .github/workflows/main.yaml
@@ -32,7 +32,7 @@ jobs:
         ports:
         - 5000:5000
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - name: Set up Python
       uses: actions/setup-python@v4
       with:
