+++ oras/tests/test_oras.py
@@ -85,6 +85,7 @@ def test_get_delete_tags(tmp_path, registry, credentials, target):
 
     # Test deleting not-existence tag
     assert not client.delete_tags(target, "v1-boop-boop")
+
     assert "v1" in client.delete_tags(target, "v1")
     tags = client.get_tags(target)
     assert not tags
