+++ scripts/test.sh
@@ -20,4 +20,4 @@ printf "ORAS_REGISTRY: ${ORAS_REGISTRY}\n"
 printf "ORAS_AUTH: ${ORAS_AUTH}\n"
 
 # Client (command line) tests
-pytest oras/
+pytest -xs oras/
